requirements.txt
matplotlib==3.5.2
nltk==3.7
pandas==1.4.2
Pillow==9.4.0
seaborn==0.12.2
snscrape==0.5.0.20230113
streamlit==1.18.1
wordcloud==1.8.2.2
